Fix build errors Update EntityConfiguration with domain class invariants WIP

+++ CustomerVehicleManagement.Shared/Models/Inventory/MaintenanceItemToRead.cs
@@ -5,7 +5,7 @@ namespace CustomerVehicleManagement.Shared.Models.Inventory
     public class MaintenanceItemToRead
     {
         public long Id { get; set; }
-        public long DisplayOrder { get; set; }
+        public int DisplayOrder { get; set; }
         public InventoryItemToRead Item { get; set; }
     }
 }
