+++ CustomerVehicleManagement.Shared/Models/Inventory/MaintenanceItemToWrite.cs
@@ -5,7 +5,7 @@ namespace CustomerVehicleManagement.Shared.Models.Inventory
     public class MaintenanceItemToWrite
     {
         public long Id { get; set; }
-        public long DisplayOrder { get; set; }
+        public int DisplayOrder { get; set; }
         public InventoryItemToWrite Item { get; set; }
     }
 }
